requirements.txt
django==5.1.2
djangorestframework==5.1.2
djangorestframework-simplejwt==2.9.0
drf-yasg==1.21.7
setuptools==75.1.0
mysqlclient==2.2.4
python-decouple==3.8
geopy==2.4.1